+++ src/main.h
@@ -1,7 +1,7 @@
 #ifndef MAIN_H
 #define MAIN_H
 // debug mode changes things like player health so playing through is easier
-#define DEBUG_MODE 1
+#define DEBUG_MODE 0
 
 #include "engine.h"